requirements.txt
# Core dependencies for transcription pipeline
openai>=1.40,<2.0
PyYAML>=6.0
tqdm>=4.66